mysoft_rag_chatbot/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.9.0
faiss-cpu==1.9.0.post1
sentence-transformers==2.5.1
numpy==2.2.0
python-dotenv==1.0.0